Fixed #12 - proto definitions split across multiple files via imports not supported
