+++ src/test/proto/ExampleServices.proto
@@ -20,31 +20,14 @@ syntax = "proto3";
 package com.example.grpc;
 
 option java_multiple_files = true;
-
+import "request/helloRequest.proto";
+import "response/helloResponse.proto";
 import "google/protobuf/empty.proto";
-
-enum Sentiment {
-    HAPPY = 0;
-    SLEEPY = 1;
-    ANGRY = 2;
-}
-
-message HelloRequest {
-    string name = 1;
-    int32 age = 2;
-    repeated string hobbies = 3;
-    map<string, string> bagOfTricks = 4;
-    Sentiment sentiment = 5;
-}
-
-message HelloResponse {
-    string greeting = 1;
-}
 
 // 4. service, unary request/response
 service GreetingService {
-    rpc greeting(HelloRequest) returns (HelloResponse);
-    rpc manyGreetingsOneReply(stream HelloRequest) returns (HelloResponse);
-    rpc oneGreetingManyReplies(HelloRequest) returns (stream HelloResponse);
-    rpc oneGreetingEmptyReply(HelloRequest) returns (google.protobuf.Empty);
+    rpc greeting(com.example.grpc.request.HelloRequest) returns (com.example.grpc.response.HelloResponse);
+    rpc manyGreetingsOneReply(stream com.example.grpc.request.HelloRequest) returns (com.example.grpc.response.HelloResponse);
+    rpc oneGreetingManyReplies(com.example.grpc.request.HelloRequest) returns (stream com.example.grpc.response.HelloResponse);
+    rpc oneGreetingEmptyReply(com.example.grpc.request.HelloRequest) returns (google.protobuf.Empty);
 }
