+++ src/test/proto/response/helloResponse.proto
@@ -1,0 +1,9 @@
+syntax = "proto3";
+
+package com.example.grpc.response;
+
+option java_multiple_files = true;
+
+message HelloResponse {
+  string greeting = 1;
+}
